🐛 동작 이상 수정
[Fix] 메서드명 변경, 코드를 컨벤션에 맞는 패키지로 이동, 어노테이션 추가 등의 수정을 거쳐 정상적으로 실행되도록 함
